+++ __init__.py
@@ -1,7 +1,7 @@
 bl_info = {
     "name": "Hyperfy Tools",
     "author": "Hyperfy",
-    "version": (1, 3, 3),
+    "version": (1, 3, 4),
     "blender": (2, 80, 0),
     "location": "View3D > Sidebar > Hyperfy",
     "description": "Tools for preparing models for Hyperfy",